duplicates are harder
Finding the duplicates worked but isolating they did not. Expanded and corrected the finding of the duplicates and did a safety comparison that the two methods returned the same results.

+++ tools/stroke.py
@@ -244,24 +244,46 @@ def targets(cursor, dry, todo: [str]):
 
 def unique(cursor, dry):
     '''remove duplicate rows in the primary table keeping the largest PK'''
-    cursor.execute(
-        'SELECT pk,COUNT(*) FROM Prime GROUP BY '
+    cursor.execute (
+        'SELECT COUNT(*) FROM Prime GROUP BY '
         'run_ID, tn_ID, task_ID, alg_ID, sv_ID, val_ID '
-        'HAVING COUNT(*) > 1;'
-    )
-    duplicates = cursor.fetchall()
-    total = sum(d[1] - 1 for d in duplicates)
+        'HAVING COUNT(*) > 1;')
+    basis = cursor.fetchall()
+    total = sum(b[-1] - 1 for b in basis)
     print(
-        f'INFO: found {len(duplicates)} duplicate groups and a total of {total} extra entries(rows)'
+        f'INFO: found {len(basis)} duplicate groups and a total of {total} extra entries(rows)'
     )
-    if not dry:
-        for group, _count in duplicates:
-            group.sort()
-            cursor.execute(
-                'DELETE FROM Prime WHERE pk = ANY(%s);', [group[:-1]]
-            )
-        cursor.commit()
-    pass
+    if total:
+        cursor.execute(
+            'SELECT pk, run_ID, task_ID, tn_ID, alg_ID, sv_ID, val_ID '
+            'FROM ( '
+            'SELECT pk, run_ID, task_ID, tn_ID, alg_ID, sv_ID, val_ID, COUNT(*)  '
+            'OVER (PARTITION BY run_ID, task_ID, tn_ID, alg_ID, sv_ID, val_ID) '
+            'AS occurs '
+            'FROM prime) '
+            'AS counted_rows '
+            'WHERE occurs > 1;'
+        )
+        rows = cursor.fetchall()
+        duplicates = {}
+        for row in rows:
+            key = row[1:]
+            if key not in duplicates: duplicates[key] = []
+            duplicates[key].append(row[0])
+        tot = sum(len(group) - 1 for group in duplicates.values())
+        if len(duplicates) != len(basis):
+            print (f'ERROR: expected {len(basis)} groupings but resolved {len(duplicates)}')
+            return
+        if total != tot:
+            print(f'ERROR: expected {total} duplicates but found {tot}')
+            return
+        if not dry:
+            for group in duplicates:
+                group.sort()
+                cursor.execute(
+                    'DELETE FROM Prime WHERE pk = ANY(%s);', [group[:-1]]
+                )
+            cursor.commit()
 
 
 if __name__ == '__main__':
